@@ -18,20 +18,20 @@
   ],
   "main": "index.js",
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=6"
   },
   "scripts": {
     "test": "mocha"
   },
   "dependencies": {
-    "through2": "^2.0.3"
+    "through2": "^3.0.0"
   },
   "devDependencies": {
     "assemble-core": "^0.31.0",
-    "delete": "^1.0.1",
-    "gulp-format-md": "^0.1.12",
-    "mocha": "^3.4.2",
-    "vinyl": "^2.0.2"
+    "delete": "^1.1.0",
+    "gulp-format-md": "^2.0.0",
+    "mocha": "^5.2.0",
+    "vinyl": "^2.2.0"
   },
   "keywords": [
     "assemble",
