@@ -1,15 +1,14 @@
 'use strict';
 
 require('mocha');
-var path = require('path');
-var assert = require('assert');
-var File = require('vinyl');
-var del = require('delete');
-var App = require('assemble-core');
-var handle = require('..');
-var app;
-
-var actual = path.resolve.bind(path, __dirname, 'actual');
+const path = require('path');
+const assert = require('assert');
+const App = require('assemble-core');
+const del = require('delete');
+const File = require('vinyl');
+const handle = require('..');
+const actual = path.resolve.bind(path, __dirname, 'actual');
+let app;
 
 describe('assemble-handle', function() {
   beforeEach(function() {
@@ -27,14 +26,14 @@ describe('assemble-handle', function() {
   it('should use a custom middleware handler', function(cb) {
     app.handler('onFoo');
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.contents = new Buffer(file.path);
+      file.contents = Buffer.from(file.path);
       next();
     });
 
-    var files = [];
-    var paths = [];
+    let files = [];
+    let paths = [];
 
-    app.src('fixtures/*.hbs', {cwd: __dirname})
+    app.src('fixtures/*.hbs', { cwd: __dirname })
       .on('data', function(file) {
         paths.push(file.path);
       })
@@ -45,31 +44,31 @@ describe('assemble-handle', function() {
       .pipe(app.dest(actual()))
       .on('finish', function() {
         assert.equal(files[0].contents.toString(), paths[0]);
-        cb()
+        cb();
       });
   });
 
   it('should handle multiple middleware for a custom handler', function(cb) {
     app.handler('onFoo');
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'a';
+      file.contents = Buffer.from(file.contents.toString() + 'a');
       next();
     });
 
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'b';
+      file.contents = Buffer.from(file.contents.toString() + 'b');
       next();
     });
 
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'c';
+      file.contents = Buffer.from(file.contents.toString() + 'c');
       next();
     });
 
-    var files = [];
-    var paths = [];
+    let files = [];
+    let paths = [];
 
-    app.src('fixtures/*.hbs', {cwd: __dirname})
+    app.src('fixtures/*.hbs', { cwd: __dirname })
       .on('data', function(file) {
         paths.push(file.path);
       })
@@ -78,37 +77,37 @@ describe('assemble-handle', function() {
         files.push(file);
       })
       .on('end', function() {
-        assert.equal(files[0].content, 'abc');
+        assert.equal(files[0].contents.toString(), 'abc');
         cb();
       });
   });
 
   it('should handle the same middleware method multiple times', function(cb) {
-    var count = 0;
+    let count = 0;
 
     app.handler('onFoo');
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'a';
+      file.contents = Buffer.from(file.contents.toString() + 'a');
       count++;
       next();
     });
 
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'b';
+      file.contents = Buffer.from(file.contents.toString() + 'b');
       count++;
       next();
     });
 
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'c';
+      file.contents = Buffer.from(file.contents.toString() + 'c');
       count++;
       next();
     });
 
-    var files = [];
-    var paths = [];
+    let files = [];
+    let paths = [];
 
-    app.src('fixtures/foo.hbs', {cwd: __dirname})
+    app.src('fixtures/foo.hbs', { cwd: __dirname })
       .on('data', function(file) {
         paths.push(file.path);
       })
@@ -122,38 +121,38 @@ describe('assemble-handle', function() {
         files.push(file);
       })
       .on('end', function() {
-        assert.equal(files[0].content, 'abcabcabcabcabcabc');
+        assert.equal(files[0].contents.toString(), 'abcabcabcabcabcabc');
         assert.equal(count, 18);
         cb();
       });
   });
 
   it('should only run a handler once when `handle.once` is used', function(cb) {
-    var count = 0;
+    let count = 0;
 
     app.handler('onFoo');
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'a';
+      file.contents = Buffer.from(file.contents.toString() + 'a');
       count++;
       next();
     });
 
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'b';
+      file.contents = Buffer.from(file.contents.toString() + 'b');
       count++;
       next();
     });
 
     app.onFoo(/\.hbs$/, function(file, next) {
-      file.content += 'c';
+      file.contents = Buffer.from(file.contents.toString() + 'c');
       count++;
       next();
     });
 
-    var files = [];
-    var paths = [];
+    let files = [];
+    let paths = [];
 
-    app.src('fixtures/foo.hbs', {cwd: __dirname})
+    app.src('fixtures/foo.hbs', { cwd: __dirname })
       .on('data', function(file) {
         paths.push(file.path);
       })
@@ -167,20 +166,20 @@ describe('assemble-handle', function() {
         files.push(file);
       })
       .on('end', function() {
-        assert.equal(files[0].content, 'abc');
+        assert.equal(files[0].contents.toString(), 'abc');
         assert.equal(count, 3);
         cb();
       });
   });
 
   it('should handle onLoad', function(cb) {
-    var count = 0;
+    let count = 0;
     app.onLoad(/./, function(file, next) {
       count++;
       next();
     });
 
-    app.src('fixtures/bar.hbs', {cwd: __dirname})
+    app.src('fixtures/bar.hbs', { cwd: __dirname })
       .pipe(app.dest(actual('out-fixtures')))
       .on('end', function() {
         assert.equal(count, 1);
@@ -189,24 +188,24 @@ describe('assemble-handle', function() {
   });
 
   it('should handle preWrite', function(cb) {
-    var count = 0;
+    let count = 0;
     app.preWrite(/./, function(file, next) {
       count++;
       next();
     });
 
-    var srcPath = path.join(__dirname, 'fixtures/foo.hbs');
-    var stream = app.dest(actual('out-fixtures'));
+    let srcPath = path.join(__dirname, 'fixtures/foo.hbs');
+    let stream = app.dest(actual('out-fixtures'));
 
     stream.once('finish', function() {
       assert.equal(count, 1);
       cb();
     });
 
-    var file = new File({
+    let file = new File({
       path: srcPath,
       cwd: __dirname,
-      contents: new Buffer("1234567890")
+      contents: Buffer.from('1234567890')
     });
     file.options = {};
 
@@ -215,24 +214,24 @@ describe('assemble-handle', function() {
   });
 
   it('should handle postWrite', function(cb) {
-    var count = 0;
+    let count = 0;
     app.postWrite(/./, function(file, next) {
       count++;
       next();
     });
 
-    var srcPath = path.join(__dirname, 'fixtures/bar.hbs');
-    var stream = app.dest(actual('out-fixtures'));
+    let srcPath = path.join(__dirname, 'fixtures/bar.hbs');
+    let stream = app.dest(actual('out-fixtures'));
 
     stream.once('finish', function() {
       assert.equal(count, 1);
       cb();
     });
 
-    var file = new File({
+    let file = new File({
       path: srcPath,
       cwd: __dirname,
-      contents: new Buffer("1234567890")
+      contents: Buffer.from('1234567890')
     });
     file.options = {};
 
